services/user-mutator/tools/requirements.txt
argparse
jinja2
PyYAML
kubernetes
